requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-ollama
ipykernel
requests
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
langserve
langchain_groq
sqlite3